+++ lib/src/spot/finder_interop.dart
@@ -2,7 +2,6 @@ import 'package:dartx/dartx.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:spot/spot.dart';
-import 'package:spot/src/spot/parent_element_filter.dart';
 
 /// Extensions that convert from the [Finder] API to spot
 extension FinderToSpot on Finder {
