+++ test/spot/exists_once_test.dart
@@ -3,7 +3,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:spot/spot.dart';
-import 'package:spot/src/spot/parent_element_filter.dart';
 
 import '../util/assert_error.dart';
